# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "UTM DSS/USS API",
    "description" : "Interface to Discovery and Synchronization Service (DSS) and UAS Service Suppliers (USS) used by participating clients to discover and interoperate.\n\nUnless otherwise specified, fields specified in a message but not declared in the API shall be ignored.",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "Authority" : [ "utm.read.operations", "utm.write.operations", "utm.read.constraints", "utm.write.constraints", "utm.read.subscriptions", "utm.write.subscriptions" ]
  } ],
...
  "paths" : {
...
    "/v1/uss/operations/{id}" : {
      "summary" : "Push/Query detailed information of an Operation.",
      "get" : {
        "tags" : [ "USS Peer-to-Peer" ],
        "summary" : "/uss/operations/{id}",
        "description" : "Retrieve full details of an Operation (directly from the entity owner).",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID for this Operation (matches Operation identifier).",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The full Operation information was retrieved successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetOperationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "One or more input parameters were missing or invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The requested Entity could not be found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.read.operations" ]
        } ]
      },
      "post" : {
        "tags" : [ "USS Peer-to-Peer" ],
        "summary" : "/uss/operations/{id}",
        "description" : "Push details of an Operation directly to a peer USS (usually because interaction with a DSS triggered this action).",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID for this Operation (matches Operation identifier).",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostOperationNotificationParameters"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "New or updated full Operation information received successfully."
          },
          "400" : {
            "description" : "* One or more parameters were missing or invalid.\n* The Entity could not be parsed, or contains illegal data.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "* The access token was decoded successfully but did not include a scope appropriate to this endpoint.\n* The client identified in the access token is not the owner of this Entity according to the receiving client's records.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The Entity version specified in this message is lower than a previously-received notification, or identical to a previously-received notification and the Entity is different.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.write.operations" ]
        } ]
      }
    },
    "/v1/uss/constraints/{id}" : {
      "summary" : "Push/Query detailed information of a Constraint.",
      "get" : {
        "tags" : [ "USS Peer-to-Peer" ],
        "summary" : "/uss/constraints/{id}",
        "description" : "Retrieve the full details of a Constraint (from the entity owner).",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID of the Constraint.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The full Constraint information was retrieved successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetConstraintResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "One or more input parameters were missing or invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The requested Entity could not be found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.read.constraints" ]
        } ]
      },
      "post" : {
        "tags" : [ "USS Peer-to-Peer" ],
        "summary" : "/uss/constraints/{id}",
        "description" : "Publish new information about a Constraint to the host of this endpoint (usually because of instructions from the DSS triggered by a subscription).",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID of the Constraint.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostConstraintNotificationParameters"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "New or updated full Constraint information received successfully."
          },
          "400" : {
            "description" : "* One or more parameters were missing or invalid.\n* The Entity could not be parsed, or contains illegal data.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "* The access token was decoded successfully but did not include a scope appropriate to this endpoint.\n* The client identified in the access token is not the owner of this Entity according to the receiving client's records.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The Entity version specified in this message is lower than a previously-received notification, or identical to a previously-received notification and the Entity is different.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.write.constraints" ]
        } ]
      }
    },
    "/v1/uss/feedback/{id}" : {
      "summary" : "Endpoint to provide feedback (errors, etc.) that might otherwise go unnoticed by this USS.",
      "put" : {
        "tags" : [ "USS Peer-to-Peer" ],
        "summary" : "/uss/feedback/{id}",
        "description" : "Notify USS of an error encountered that might otherwise go unnoticed.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the feedback.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UUIDv4"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PutFeedbackParameters"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Feedback received successfully."
          },
          "400" : {
            "description" : "* One or more parameters were missing or invalid.\n* The feedback could not be parsed, or contains illegal data.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FeedbackURL"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "* The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The Entity version specified in this message is lower than a previously-received notification, or identical to a previously-received notification and the Entity is different.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.read.operations", "utm.write.operations", "utm.read.constraints", "utm.write.constraints" ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "EntityUUID" : {
        "description" : "Universally-unique identifier for an Entity communicated through the DSS.  Formatted as UUIDv4.",
        "anyOf" : [ {
          "$ref" : "#/components/schemas/UUIDv4"
        } ]
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Human-readable message indicating what error occurred and/or why."
          }
        },
        "description" : "Data provided when an off-nominal condition was encountered."
      },
...
    },
    "securitySchemes" : {
      "Authority" : {
        "type" : "oauth2",
        "description" : "Authorization from, or on behalf of, an authorization authority.  This authority shall issue access tokens that are JSON Web Tokens as defined in RFC 7519, using the `RS256` algorithm for the signature, and publish to all providers the public key for verifying that signature.\n\nThe following fields shall be included in the JWT claim for access tokens issued by this authority:\n* `iss`, with the URL at which the token generation request was received.\n* `exp`, with a time no further than 1 hour in the future.\n* `sub`, with unique ID of the client requesting the access token.\n* `scope`, with an array of strings indicating the scopes granted.\n\nMultiple scopes for a single token shall be granted if requested.  The tokens granted by this authority shall protect against reuse of received tokens to impersonate the sender to other recipients (via use of the `aud` claim or other means).\n\nClients shall provide these access tokens in an `Authorization` header in the form `Bearer <token>` in accordance with RFC 6750.",
        "flows" : {
          "clientCredentials" : {
            "tokenUrl" : "https://token_vendor.com/oauth/token",
            "scopes" : {
              "utm.read.operations" : "Client may query, read, or create subscriptions for Operations for UTM.",
              "utm.write.operations" : "Client may create, edit, and delete Operations for UTM.",
              "utm.read.constraints" : "Client may query, read, or create subscriptions for Constraints for UTM.",
              "utm.write.constraints" : "Client may create, edit, and delete Constraints for UTM."
            }
          }
        }
      }
    }
  }
}